% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ConfSVM.R
\docType{package}
\name{ConfSVM}
\alias{ConfSVM}
\alias{ConfSVM-package}
\title{ConfSVM: A Package for several Ensemble Support Vector Machine Models}
\description{
The ConfSVM package contains three differenct Ensemble SVM models,
they all train SVM models on subset of data and combine the results together.
The three models are from the following three papers:
}
\details{
1. Gu, Q., & Han, J. (2013). Clustered support vector machines. In proceedings of the sixteenth international conference on artificial intelligence and statistics (pp. 307-315).
2. Hsieh, C. J., Si, S., & Dhillon, I. S. (2013). A divide-and-conquer solver for kernel support vector machines. arXiv preprint arXiv:1311.0914.
3. Collobert, R., Bengio, S., & Bengio, Y. (2002). A parallel mixture of SVMs for very large scale problems. Neural computation, 14(5), 1105-1114.
}
\section{ConfSVM functions}{

clusterSVM
dcSVM
gaterSVM
}

