% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/confSVMPredict.R
\name{confSVMPredict}
\alias{confSVMPredict}
\title{predic function for convsvm}
\usage{
confSVMPredict(model = "1999", gamma = 3.125, cost = 1, train.x = NULL,
  train.y = NULL, ...)
}
\arguments{
\item{model}{model to use}
}
\description{
predic function for convsvm
}

