% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/confSVMTrain.R
\name{confSVMTrain}
\alias{confSVMTrain}
\title{training function for convsvm}
\usage{
confSVMTrain(model = "1999", gamma = 3.125, cost = 1, train.x = NULL,
  train.y = NULL, kappa = 0.1, tau = 0.1, ...)
}
\arguments{
\item{model}{model to use}
}
\description{
training function for convsvm
}

